Use debouncedHighlight for small multiples
