+++ src/modes/single/SingleLocation.svelte
@@ -11,30 +11,8 @@
   } from '../../stores';
   import { sensorList } from '../../stores/constants';
   import SensorCard from './SensorCard.svelte';
-  import debounce from 'lodash-es/debounce';
   import { selectionColors } from '../../theme';
-  import { resolveHighlightedTimeValue } from '../overview/vegaSpec';
-
-  let highlightTimeValue = null;
-
-  // Each mouse move across the chart results in a mouseout for previous highlight
-  // date immediately followed by mouseover for the next date, if any.  We need to
-  // ignore the first event unless there is no second.  A very short debounce time
-  // achieves that goal, at least most of the time.
-  const debouncedHighlightTime = debounce(
-    (value) => {
-      highlightTimeValue = value;
-    },
-    1,
-    { leading: false, trailing: true },
-  );
-
-  function onHighlight(e) {
-    const value = resolveHighlightedTimeValue(e);
-    if (highlightTimeValue !== value) {
-      debouncedHighlightTime(value);
-    }
-  }
+  import { onHighlight, highlightTimeValue } from '../overview/vegaSpec';
 
   $: selectedLevels = new Set($currentMultiSelection.map((d) => d.info.level));
   function filterItem(item) {
